pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "icaf"
version = "1.0.0"
description = "ITSAR Compliance Automation Framework"
requires-python = ">=3.9"

dependencies = [
    "typer",
    "rich",
    "selenium",
    "pyyaml",
    "python-docx",
    "pyautogui"
]

[project.scripts]
icaf = "icaf.cli:main"

[tool.setuptools.packages.find]
include = ["icaf*"]
exclude = ["logs*", "output*", "myenv*"]